pyproject.toml
[tool.poetry]
name = "hydra-project"
version = "0.1.0"
description = ""
authors = ["jihoahn9303 <rediate03@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
hydra-core = "^1.3.2"
torch = {url = "https://download.pytorch.org/whl/cu116/torch-1.12.1%2Bcu116-cp39-cp39-linux_x86_64.whl"}
torchvision = {url = "https://download.pytorch.org/whl/cu116/torchvision-0.13.1%2Bcu116-cp39-cp39-linux_x86_64.whl"}
tensorboard = "^2.15.1"
pytorch-lightning = "^1.7.7"
pydantic = "^2.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
